image-dedup/requirements.txt
Flask==2.3.3
Pillow==10.0.0
imagehash==4.3.1
requests==2.31.0
gunicorn==21.2.0
